apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "studioops-api"
version = "0.1.0"
description = "StudioOps AI API Service"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.1",
    "python-dotenv>=1.0.0",
    "openai>=1.3.9",
    "langfuse>=2.0.0",
    "chromadb>=0.4.22",
    "trello>=0.9.1",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1",
    "email-validator>=2.0.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "flower>=2.0.1",
    "prometheus-client>=0.19.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-playwright>=0.3.3",
    "playwright>=1.39.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "tox>=4.11.3",
    "coverage>=7.3.2",
    "faker>=19.6.2",
    "factory-boy>=3.2.1",
    "freezegun>=1.2.2",
    "respx>=0.20.2",
    "fastapi-testclient>=0.2.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=apps/api --cov-report=term-missing --cov-report=html --cov-report=xml"

[tool.coverage.run]
source = ["apps/api"]
omit = [
    "apps/api/tests/**",
    "apps/api/__pycache__/**",
    "apps/api/**/__pycache__/**",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "B", "C", "N", "R", "T", "Q", "S", "PL", "RUF"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["S101", "S102", "S103"]

[tool.black]
line-length = 88
target-version = ["py39"]
skip-string-normalization = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true

[project.scripts]
api = "apps.api.main:app"
dev = "uvicorn apps.api.main:app --reload --host 0.0.0.0 --port 8000"
test = "pytest tests/ -v"
test-cov = "pytest tests/ --cov=apps/api --cov-report=html"
lint = "ruff check apps/api/ tests/"
format = "black apps/api/ tests/"
typecheck = "mypy apps/api/"
setup-test-db = "python apps/api/tests/setup_test_db.py"